% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFuSIM.R
\name{cFuSIM}
\alias{cFuSIM}
\title{This function computes the coeffcients of the index function}
\usage{
cFuSIM(y, xfd, spline_basis, threshold = 1e-05, maxit = 150,
  lambda = 10000)
}
\arguments{
\item{y}{a vector of response.}

\item{xfd}{a fda object contains the functional predictor}

\item{spline_basis}{the B-spline basis for the index function}

\item{lambda}{a positive number, the sparsity parameter}
}
\description{
This function computes the coeffcients of the index function
}
\examples{
\dontrun{
data(binary_dat);xmat = binary_dat$x;y=binary_dat$y
res = sfpcs_binary(xmat,y,npc_select=2,theta=1,xmat_new = xmat)
res$fitted
mean(y==res$fitted) # fitted accuracy
res$fitted # prediction on new data 
res$sfpcs # sFPCs 
res$beta_fd # coefficient function
}

}
