% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFuSIM.R
\name{cFuSIM_index}
\alias{cFuSIM_index}
\title{This function computes the coeffcients of the index function}
\usage{
cFuSIM_index(y, xfd, spline_basis, threshold = 1e-05, maxit = 150,
  lambda = 10000)
}
\arguments{
\item{y}{a vector of response.}

\item{xfd}{a fda object contains the functional predictor}

\item{spline_basis}{the B-spline basis for the index function}

\item{lambda}{a positive number, the sparsity parameter}
}
\description{
This function computes the coeffcients of the index function
}
\examples{
\dontrun{
library(fda)
library(cFuSIM)
data(bike_cFuSIM)
norder=4 ## cubic B-spline
nbasis=norder+length(timepts)-2; 
spline.basis=create.bspline.basis(rangeval=c(1,24),nbasis,norder,timepts)
wull = bike$temp
xfds=  Data2fd(y=wull\%>\%t, argvals=bike$timepts)
y = bike$y
train_sample = 1:length(y)
y = y[train_sample]
xfd = xfds[train_sample]
res_c = cFuSIM(y, xfd, spline_basis)
beta_fd = fd(res_c$coefBeta, res_c$basisBeta)
plot(beta_fd)
}
}
