% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfda.R
\name{fdagg}
\alias{fdagg}
\title{This function computes the coeffcients of the index function}
\usage{
fdagg(fdobj, ylab = "Value", xlab = "Time", addzero = FALSE)
}
\arguments{
\item{fdobj}{fda object}
}
\description{
This function computes the coeffcients of the index function
}
\examples{
\dontrun{
library(fda)
library(cFuSIM)
spline.basis=create.bspline.basis(rangeval=c(1,24),nbasis=10,norder)
beta_fd = fd(rnorm(10), spline.basis)
plot(beta_fd)
}
}
